pyproject.toml
[tool.poetry]
name = "bs4_corona"
version = "0.1.0"
description = ""
authors = ["Andre"]

[tool.poetry.dependencies]
python = "^3.8"
bs4 = "^0.0.1"
requests = "^2.23.0"
sqlalchemy = "^1.3.17"
fastapi = "^0.54.2"
uvicorn = "^0.11.5"
python-dotenv = "^0.13.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.1"
pylint = "^2.5.2"
yapf = "^0.30.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
